requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
psycopg[binary]==3.3.3
pytest==8.4.1
httpx==0.28.1
python-dotenv==1.1.1
